Refine 2 datasets
Only three variables exist in the candidate data that was not in the polling data. Those were added to the polling data to create a fuller dataset. Also removed all instances where there were more than three people in the race. Also removed all instance of Louisiana. Also added a binary indicator for if a candidate won or not